C++ Generation: get rid of "std::reference_wrapper" usage
Motivation: while the usage of std::reference_wrapper for the API is more clean wrt the lifecycle semantics, currently it is much easier to keep things simple and use std::shared_ptr for every reference signature. Later when the C++ simulator is working, the replacement with more C++ conform ways like the reference_wrapper are optional extensions for this project.