C++ Generation: for HamsterGame generate static factory method
Motivation: while constructing the HamsterGame it has to pass itself to the Territory, which is also constructed at this time. But while the constructor of HamsterGame is not left, it is illegal to access "shared_from_this". As a solution use a static "create" method which solves this issue. This is currently only neccessary for "MiniProgrammingWorld"-instances.